// src/app/interfaces/segurity-interfaces/user.interface.ts
// Generated by https://quicktype.io
export interface IUserByIdResponse {
  data:  IUser;
}
// Generated by https://quicktype.io

export interface IUserListResponse {
  data:  IUser[];
  total: number;
}

export interface IUser {
  status:       boolean;
  createAt:     string;
  id:           string;
  name:         string;
  surname:      string;
  fullname:     string;
  address:      null | string;
  cityname:     null | string;
  email:        string;
  phone:        null | string;
  countryCode:  string;
  postalcode:   string;
  campusId:     null;
  roles:        Role[];
  organization: Organization;
  customertype: Customertype | null;
  timezone:     Timezone | null;
}

export interface Customertype {
  status:       boolean;
  userCreate:   null;
  createAt:     string;
  id:           string;
  customerType: string;
}

export interface Organization {
  status:        boolean;
  userCreate:    string;
  createAt:      string;
  id:            string;
  bussinnesName: string;
  document:      string;
  address:       string;
  email:         string;
  phone:         string;
  costPerMinute: string;
  logoUrl:       string;
  backdropUrl:   string;
  campus:        Campus[];
}

export interface Campus {
  status:     boolean;
  userCreate: string;
  createAt:   string;
  id:         string;
  campusName: string;
  city:       string;
  address:    null;
}

export interface Role {
  id:   string;
  name: string;
}

export interface Timezone {
  status:       boolean;
  userCreate:   null;
  createAt:     null;
  id:           string;
  zoneName:     string;
  countryCode:  string;
  abbreviation: string;
  timeStart:    string;
  gmtOffset:    number;
  dst:          string;
}